clearway-for-whatsapp/whatsapp_bridge.py
import os
import signal
import time
from pathlib import Path
from dataclasses import dataclass
from typing import List, Dict, Optional, Any
from datetime import datetime
from neonize.client import NewClient
from neonize.events import ConnectedEv, MessageEv, PairStatusEv
from neonize.types import JID
from neonize.utils import log
from dotenv import load_dotenv

load_dotenv()

# --- Data Models ---

@dataclass
class UserProfile:
    id: str
    display_name: str
    push_name: Optional[str]

@dataclass
class Message:
    id: str
    sender_id: str
    sender_name: str
    content: str
    timestamp: datetime
    conversation_id: str
    is_unread: bool = False

@dataclass
class Conversation:
    id: str
    title: str
    last_message_time: datetime
    messages: List[Message]
    unread_count: int = 0
    thread_type: str = "Chat" # Chat, Group, Broadcast

# --- Extractor Class ---

class WhatsAppExtractor:
    def __init__(self, db_path: str = "whatsapp.db"):
        self.client = NewClient(db_path)
        self.conversations: List[Conversation] = []
        self.is_connected = False

    def on_connected(self, client: NewClient, event: ConnectedEv):
        log.info("Connected to WhatsApp")
        self.is_connected = True

    def __enter__(self):
        self.client.event_handler.register(ConnectedEv, self.on_connected)
        self.client.connect()
        # Wait for connection (simple sync wait for CLI demo)
        timeout = 30
        start = time.time()
        while not self.is_connected and time.time() - start < timeout:
            time.sleep(1)
        return self

    def __exit__(self, exc_type, exc_val, exc_tb):
        self.client.disconnect()

    def get_conversations(self) -> List[Conversation]:
        # Note: Neonize/Whatsmeow typically works via event streams. 
        # For a "snapshot" bridge, we'd query the local store if available, 
        # but whatsmeow store is usually internal.
        # This implementation shows how we'd map the data once received/synced.
        
        # In a real bridge, we'd use client.get_contact, client.get_group_info, etc.
        # For now, this is a skeleton showing the interface compatibility.
        
        log.info("Fetching conversations...")
        # Placeholder for actual store query logic
        return self.conversations

# --- Execution ---

def main():
    print(f"Initializing WhatsApp Message Bridge...")
    try:
        # Note: WhatsApp requires QR pairing on first run.
        # Neonize will print the QR code to terminal or save it.
        with WhatsAppExtractor() as extractor:
            # Short wait to allow some sync
            print("Waiting for sync...")
            time.sleep(5)
            
            conversations = extractor.get_conversations()
            print(f"\nFound {len(conversations)} conversations sync'd so far.\n")
            
            if not conversations:
               print("No conversations found yet. If this is a new run, please ensure you've paired the device.")

    except Exception as e:
        print(f"Fatal error: {e}")
        import traceback
        traceback.print_exc()

if __name__ == "__main__":
    main()
